new directory for good run list creation
